+++ application/views/Homepage.php
@@ -2,26 +2,32 @@
 	<div class="container">
 		<div class="row">
 			<div class="col-md-8" style = "margin-right: 15px;">
+
+                <!-- Generates the row of 4 stocks -->
 				<?php for ($row = 0; $row < count($stockInfoArray); $row++){ ?>
 					<div class = "row div-stock-dash-wrapper padding-top-5">
+
+                        <!-- Generates the individual stocks in each row -->
 						<?php for ($col = 0; $col < count($stockInfoArray[$row]); $col++ ){ ?>
-							<div class = "col-md-3 norm">
-								<div class= "row">
-									<div class = "col-md-12">
-										<h2><?php echo $stockInfoArray[$row][$col]['Code'] ?></h2>
-										<span class = "span-stock-dash-subtitle"><?php echo $stockInfoArray[$row][$col]['Name'] ?></span>
-										<div class= "row">
-											<div class = "col-md-6">
-												<h1><?php echo $stockInfoArray[$row][$col]['Value'] ?></h1>
-											</div>
-											<div class = "col-md-6 text-right">
-												<span><?php echo $stockInfoArray[$row][$col]['Category'] ?></span>
-												<span><?php echo $stockInfoArray[$row][$col]['Value'] ?></span>
-											</div>
-										</div>
-									</div>
-								</div>
-							</div>
+                            <div class = "col-md-3 norm">
+                                <a href="/stockhistory?selectedStock=<?php echo $stockInfoArray[$row][$col]['Code'] ?>">
+                                    <div class= "row">
+                                        <div class = "col-md-12">
+                                            <h2><?php echo $stockInfoArray[$row][$col]['Code'] ?></h2>
+                                            <span class = "span-stock-dash-subtitle"><?php echo $stockInfoArray[$row][$col]['Name'] ?></span>
+                                            <div class= "row">
+                                                <div class = "col-md-6">
+                                                    <h1><?php echo $stockInfoArray[$row][$col]['Value'] ?></h1>
+                                                </div>
+                                                <div class = "col-md-6 text-right">
+                                                    <span><?php echo $stockInfoArray[$row][$col]['Category'] ?></span>
+                                                    <span><?php echo $stockInfoArray[$row][$col]['Value'] ?></span>
+                                                </div>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </a>
+                            </div>
 						<?php } ?>
 					</div>
 				<?php } ?>
@@ -42,7 +48,6 @@
 					</div>
 				</div>
 				{/PlayerInfo}
-				<!-- <a href="#menu-toggle" class="btn btn-default" id="menu-toggle">Toggle Menu </a> -->
 			</div>
 		</div>
 	</div>